+++ dataprocessing.h
@@ -23,6 +23,7 @@ struct SimpleGraphData{
     vector d_pitch_log;
     vector d_intensive_original;
     vector d_intensive;
+    vector d_intensive_norm;
     vector d_spec_proc;
     vector d_spec;
     vector d_cepstrum;