+++ defines.h
@@ -37,7 +37,7 @@
 #define NORM_FROM 0.0
 #define NORM_TO   1.0
 
-#define APP_VER "0.2.8"
+#define APP_VER "0.2.9"
 #define FULL_APP_VERIOSN ("IntonTrainer (ver. " + QString(APP_VER) + " - " + QString(__DATE__) + ")")
 
 #endif // DEFINES